Custom Responses/Headers

+++ db/db_user.py
@@ -2,6 +2,8 @@ from sqlalchemy.orm.session import Session
 from schemas import UserBase
 from db.models import DbUser
 from db.hash import Hash
+from fastapi import HTTPException, status
+
 
 def create_user(db: Session, request: UserBase):
     new_user = DbUser(
@@ -20,12 +22,19 @@ def get_all_users(db: Session):
     return db.query(DbUser).all()
 
 def get_user(db: Session, id: int):
+    user = db.query(DbUser).filter(DbUser.id == id).first()
     # Handle any exceptions
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {id} not found")
+    return 
 
 def update_user(db: Session, id: int, request: UserBase):
-    # Handle any exceptions
     user = db.query(DbUser).filter(DbUser.id == id)
+    # Handle any exceptions
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {id} not found")
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -35,8 +44,11 @@ def update_user(db: Session, id: int, request: UserBase):
     return "Okay"
 
 def delete_user(db: Session, id: int):
-    # Handle any exceptions
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    # Handle any exceptions
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {id} not found")
     db.delete(user)
     db.commit()
     return 'Ok deleted'
